+++ src/main/java/com/example/syncrowbackend/common/exception/ErrorCode.java
@@ -11,6 +11,7 @@ public enum ErrorCode{
     POST_NOT_FOUND_ERROR(400, "POST-NOT-FOUND", "POST NOT FOUND ERROR"),
     POST_NOT_DELETABLE(400, "POST_NOT_DELETABLE", "POST NOT DELETABLE"),
     GROUP_NOT_FOUND_ERROR(400, "GROUP-NOT-FOUND", "GROUP NOT FOUND ERROR"),
+    DUPLICATED_GROUP_ENTER(400, "DUPLICATED_GROUP_ENTER", "DUPLICATED GROUP ENTER"),
     FRIEND_REQUEST_ROOM_NOT_FOUND_ERROR(400, "FRIEND_REQUEST_ROOM_NOT_FOUND_ERROR", "FRIEND_REQUEST_ROOM NOT FOUND ERROR"),
     PERMISSION_NOT_GRANTED_ERROR(403, "PERMISSION-NOT-GRANTED", "PERMISSION NOT GRANTED ERROR"),
     INTERNAL_SERVER_ERROR(500, "INTERNAL-SERVER-ERROR", "INTERNAL SERVER ERROR"),
