+++ src/main/java/com/example/syncrowbackend/friend/entity/Notification.java
@@ -2,7 +2,7 @@ package com.example.syncrowbackend.friend.entity;
 
 import com.example.syncrowbackend.common.entity.BaseTimeEntity;
 import com.example.syncrowbackend.friend.enums.NotificationStatus;
-import com.example.syncrowbackend.user.entity.User;
+import com.example.syncrowbackend.auth.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
