+++ src/main/java/com/example/syncrowbackend/friend/repository/FriendRequestRepository.java
@@ -3,7 +3,7 @@ package com.example.syncrowbackend.friend.repository;
 import com.example.syncrowbackend.friend.entity.FriendRequest;
 import com.example.syncrowbackend.friend.entity.Post;
 import com.example.syncrowbackend.friend.enums.FriendRequestStatus;
-import com.example.syncrowbackend.user.entity.User;
+import com.example.syncrowbackend.auth.entity.User;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
