+++ src/main/java/com/example/syncrowbackend/friend/repository/NotificationRepository.java
@@ -3,7 +3,7 @@ package com.example.syncrowbackend.friend.repository;
 import com.example.syncrowbackend.friend.entity.FriendRequest;
 import com.example.syncrowbackend.friend.entity.Notification;
 import com.example.syncrowbackend.friend.enums.NotificationStatus;
-import com.example.syncrowbackend.user.entity.User;
+import com.example.syncrowbackend.auth.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
