+++ src/main/java/com/example/syncrowbackend/friend/service/FriendService.java
@@ -3,7 +3,7 @@ package com.example.syncrowbackend.friend.service;
 import com.example.syncrowbackend.friend.dto.FriendReactDto;
 import com.example.syncrowbackend.friend.dto.FriendRequestDto;
 import com.example.syncrowbackend.friend.enums.FriendReaction;
-import com.example.syncrowbackend.user.entity.User;
+import com.example.syncrowbackend.auth.entity.User;
 
 public interface FriendService {
     void friendRequest(FriendRequestDto requestDto, User user);
